+ ansrole: add ability to run on inventory groups
in order to apply to groups and not just single hosts, we have to be
able to run against particular ansible inventory.  this means basically
we want to pass in '--inventory' to ansible and then give an identifier
to work from in that inventory, such as a group list (like a gender, for
example).

in order to do this we add parsing code so we can take two different
kinds of "onto" args, of the form

    hostlist:role OR
    inventory:grouplist:role

so the user can give, eg:

    setup/etc/genders:some_gender:role_to_apply

for example:

    ansrole -u root etc/genders:dfw3:skel skel_user=scott

to apply my skeleton rol to all dfw3 machines