= ansrole: add comments about adjusting usage twice
unfortunately we have to duplicate usage description updates because
it's embedded both in the comment header and there's a function usagex()
which outputs it for the user.  ideally we can just parse our own
comment header to get this in usagex() but this isn't implemented yet.
in the meantime, leave a comment alerting the editor to his obligation
to edit it twice